requirements.txt
Flask==3.0.0
Pillow==9.5.0
Pillow==10.1.0
